chore: combine attribute impl
